fbgl-rs: made code changes or something idk

+++ src/text.rs
@@ -6,14 +6,14 @@ use fontdue::Font;
 
 pub struct TextRenderSettings {
     pub font: Font,
-    pub size: u16,
+    pub size: u32,
     pub blend: bool,
 }
 
 pub trait TextOperations: GraphicsRenderer {
-    fn text(&mut self, col: Self::Color, x: u16, y: u16, txt: String, opts: TextRenderSettings);
+    fn text(&mut self, col: Self::Color, x: u32, y: u32, txt: String, opts: TextRenderSettings);
 }
 
 impl<T: GraphicsRenderer> TextOperations for T {
-    fn text(&mut self, col: Self::Color, x: u16, y: u16, txt: String, opts: TextRenderSettings) {}
+    fn text(&mut self, col: Self::Color, x: u32, y: u32, txt: String, opts: TextRenderSettings) {}
 }
